+++ CHANGELOG.md
@@ -74,6 +74,8 @@
 - Removed image_colour_tint!
 - Image#invert! is now C++ backed
 - Removed image_colour_invert!
+- Image#grayscale! was renamed to Image#greyscale! and is now C++ backed
+- Removed image_colour_grayscale!
 
 ## v0.3.14.1